Added extend mode to netdevil protocol + moved ini implementation to internal/netdevil
